feat: refactor categories screen and update navigation bar
This commit refactors the categories screen, updates the navigation bar component, and removes unused files related to categories and transaction screens.

-   **Categories Screen Refactor:**
    -   Removed `CategoriesContentView.swift` and moved its content into `CategoriesScreen.swift` for better code organization.
    -   Updated `CategoriesScreen` to directly use and manage state and events within the view.
    - Update `SelectIconCategoryView` to handle icon selection.
    -   Updated `CategoriesScreen` to use `CategoriesContentView`
    - Refactor `TabContentView` in `CategoriesScreen` to use `TabView`.
    - Refactor `TabBarItemView` to use `Int32` for tab indexes.
    - Refactor `TabBarView` to use `Int32` for tab indexes.
-   -   Updated state and event handling logic for search and tab changes.
-   -   Refactor `NavigationTopAppbar` to use `onAction` instead of `navigateBack`
-  -   Refactor `transactionTopAppBarView` to use `NavigationTopAppbar`
-  - Update `CategoriesScreen` to use `NavigationTopAppbar`

-   **Navigation Bar Updates:**
    -   Refactored `NavigationTopAppbar` to use an `onAction` callback instead of `navigateBack`.
    -   Added a `showDrawer` parameter to `NavigationTopAppbar` to toggle the drawer icon.
    -   Updated all the screens to use new `NavigationTopAppbar`
    - Update `AppNavigation` to use `showDrawer` parameters for `categories` and `transactions`.
    -   Added `default` function in `CategoriesState`.
-   -   Removed `TransactionNavBarView.swift`.
-    - Refactor `TransactionTopAppBarView` to use `NavigationTopAppbar`.
- **Remove Unused Files**:
    -   Removed `CategoriesState.swift` and `TransactionNavBarView.swift` as these functionalities have been moved to Kotlin.

+++ iosApp/iosApp/designSystem/component/tab/TabBarView.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 struct TabBarView: View {
-    @Binding var currentTab: Int
+    @Binding var currentTab: Int32
     @Namespace var nameSpace
     
     var tabBarOptions: [String] = ["View 1", "View 2"]
@@ -13,7 +13,7 @@ struct TabBarView: View {
                     currentTab: $currentTab,
                     namespace: nameSpace.self,
                     tabBarItemName: name,
-                    tab: index
+                    tab: Int32(index)
                 )
             }
         }
